BPMN File
The following BPMN file shoes a generic process for how an architectural/source model can first be used with BlenderBIM or any other open source BIM/programming language to extract the properties, which will serve as an input in any FEM software to carry out Static Structural Analysis of the model. This will help validate the source model and will give an output for documentation the process.  Additionally, the process also takes into account how code validation will be performed for various structural members used, whether that they comply with material standards and EC specifications, before they're ready to be used with the FEM software.